Add ObjectWidgetUtil
